+++ src/pharmpy/modeling/odes.py
@@ -1,5 +1,6 @@
 import sympy
 
+import pharmpy.symbols
 from pharmpy.parameter import Parameter
 from pharmpy.statements import Assignment, Bolus, CompartmentalSystem, ExplicitODESystem, Infusion
 
@@ -38,15 +39,15 @@ def _have_zero_order_absorption(model):
     return False
 
 
-def absorption(model, order, rate=None):
-    """Set or change the absorption for a model
+def absorption_rate(model, order):
+    """Set or change the absorption rate of a model
 
     Parameters
     ----------
     model
         Model to set or change absorption for
     order
-        'bolus', 0 or 1
+        'instant', 'ZO', 'FO' or 'seq-ZO-FO'
     """
     statements = model.statements
     odes = statements.ode_system
@@ -54,16 +55,14 @@ def absorption(model, order, rate=None):
         raise ValueError("Setting absorption is not supported for ExplicitODESystem")
 
     depot = odes.find_depot()
-    order = str(order)
-    if order == 'bolus':
+    if order == 'instant':
         if depot:
             to_comp, _ = odes.get_compartment_flows(depot)[0]
             to_comp.dose = depot.dose
             ka = odes.get_flow(depot, odes.find_central())
             odes.remove_compartment(depot)
             symbols = ka.free_symbols
-            for s in symbols:
-                statements.remove_symbol_definition(s, odes)
+            statements.remove_symbol_definitions(symbols, odes)
             model.statements = statements
             model.remove_unused_parameters_and_rvs()
         elif _have_zero_order_absorption(model):
@@ -71,25 +70,44 @@ def absorption(model, order, rate=None):
             old_symbols = dose_comp.free_symbols
             dose_comp.dose = Bolus(dose_comp.dose.amount)
             unneeded_symbols = old_symbols - dose_comp.dose.free_symbols
-            for symb in unneeded_symbols:
-                statements.remove_symbol_definition(symb, odes)
+            statements.remove_symbol_definitions(unneeded_symbols, odes)
             model.remove_unused_parameters_and_rvs()
-    elif order == '0':
+    elif order == 'ZO':
         if not _have_zero_order_absorption(model):
-            pass
-    elif order == '1':
+            dose_comp = odes.find_dosing()
+            symbols = dose_comp.free_symbols
+            new_dose = Infusion(dose_comp.dose.amount,
+                                duration=pharmpy.symbols.symbol('MAT') * 2)
+            if depot:
+                to_comp, _ = odes.get_compartment_flows(depot)[0]
+                to_comp.dose = new_dose
+                ka = odes.get_flow(depot, odes.find_central())
+                odes.remove_compartment(depot)
+                symbols |= ka.free_symbols
+            else:
+                dose_comp.dose = new_dose
+            statements.remove_symbol_definitions(symbols, odes)
+            mat_param = Parameter('TVMAT', init=0.1, lower=0)
+            model.parameters.add(mat_param)
+            imat = Assignment('MAT', mat_param.symbol)
+            model.statements.insert(0, imat)
+            model.remove_unused_parameters_and_rvs()
+    elif order == 'FO':
         if not depot:
             dose_comp = odes.find_dosing()
             depot = odes.add_compartment('DEPOT')
             depot.dose = Bolus(dose_comp.dose.amount)
+            symbols = dose_comp.free_symbols
             dose_comp.dose = None
+            statements.remove_symbol_definitions(symbols, odes)
             mat_param = Parameter('TVMAT', init=0.1, lower=0)
             model.parameters.add(mat_param)
             imat = Assignment('MAT', mat_param.symbol)
-            model.statements = model.statements.insert(0, imat)     # FIXME: Don't set again
-            odes.add_flow(depot, dose_comp, 1 / mat_param.symbol)
+            model.statements.insert(0, imat)
+            odes.add_flow(depot, dose_comp, 1 / pharmpy.symbols.symbol('MAT'))
             model.remove_unused_parameters_and_rvs()
     else:
-        raise ValueError(f'Requested order {order} but only orders bolus, 0 and 1 are supported')
+        raise ValueError(f'Requested order {order} but only orders '
+                         f'instant, FO, ZO and seq-ZO-FO are supported')
 
     return model